用户模块

+++ resources/views/post/show.blade.php
@@ -4,12 +4,16 @@
             <div class="blog-post">
                 <div style="display:inline-flex">
                     <h2 class="blog-post-title">{{$post->title}}</h2>
-                    <a style="margin: auto"  href="/posts/{{$post->id}}/edit">
-                        <span class="glyphicon glyphicon-pencil" aria-hidden="true"></span>
-                    </a>
-                    <a style="margin: auto"  href="/posts/{{$post->id}}/delete">
-                        <span class="glyphicon glyphicon-remove" aria-hidden="true"></span>
-                    </a>
+                    @if (Auth::user()->can('update', $post))
+                        <a style="margin: auto"  href="/posts/{{$post->id}}/edit">
+                            <span class="glyphicon glyphicon-pencil" aria-hidden="true"></span>
+                        </a>
+                    @endif
+                    @if (Auth::user()->can('delete', $post))
+                        <a style="margin: auto"  href="/posts/{{$post->id}}/delete">
+                            <span class="glyphicon glyphicon-remove" aria-hidden="true"></span>
+                        </a>
+                    @endif
                 </div>
 
                 <p class="blog-post-meta">{{$post->created_at->toFormattedDateString()}} <a href="#">Kassandra Ankunding2</a></p>
